backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
aiosqlite==0.20.0
httpx==0.28.1
groq==0.13.1
google-generativeai==0.8.3
openai==1.59.3
python-multipart==0.0.20
